+++ src/providers/backend.h
@@ -24,6 +24,7 @@
 
 #include "providers/data_provider.h"
 #include "providers/fail_over.h"
+#include "providers/be_netlink.h"
 #include "providers/be_refresh.h"
 #include "providers/data_provider/dp.h"
 #include "util/child_common.h"
@@ -104,6 +105,7 @@ struct be_ctx {
     struct be_ptask *check_if_online_ptask;
 
     struct be_refresh_ctx *refresh_ctx;
+    struct be_netlink_ctx *nlctx;
 
     size_t check_online_ref_count;
     int check_online_retry_delay;
